src/backend/requirements.txt
watchdog>=3.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-frontmatter>=1.0.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
sentence-transformers>=2.2.0
numpy>=1.24.0
